Use DMA engine to transfer display buffer
Instead of having the CPU Idle while the Data is being transfered to the
Display, Use the DMA engine to transfer display buffer into the I2C
peripheral. allowing the CPU time to be used otherwise.

The use of the DMA needed to change the way the peripheral is
initialized. Most of the data initialization is done via a Macro that
sets up all the necessary data in the display struct. To initialize the
display, the ssd1306_init function still needs to be run. The define
also changes the implementation of the 'ssd1306_show' function to use
the DMA instead of 'fancy_write'
Also add commands to the CMakeLists.txt file of the example to allow for changing
the compile Define on the target.